refactor(session-persistence): extract state owner
